SurfaceView: Set _renderedSource to false only if surface update was scheduled.
The problematic scenario is:
1. New SurfaceView is created SurfaceView::_renderedSource is set to true.
2. An update event is received and queued in
   SurfaceEventListener::_queue.
3. ilixi:Engine starts executing registered callback functions,
   SurfaceEventListener::funck is called which in turn calls
   SurfaceView::onSourceUpdate.
4. In case visible() returns false we don't actually call update(), but we
   set SurfaceView::_renderedSource to false.
5. SurfaceView::onSourceUpdate returs true which means the event is
   removed from SurfaceEventListener::_queue.
6. Because we never really update()'d the SurfaceView,
   SurfaceView::paint() will not be called until surface visibility
   changes. In short SurfaceView::renderedSouce will remain false for
   as long as surface is invisible().

At this point if the SurfaceView remains invisible but we do reveive
and update event for it the compositor will enter a tight loop. The
SurfaceEventListener::_queue will contain the event so
SurfaceEventListener::funck() will keep calling
SurfaceView::onSourceUpdate. Because SurfaceView::_renderedSource is
false the function will exit almost immediately returning false. This
means the event wont be removed from SurfaceEventListener::_queue. The
end result is the compositor will keep using 100% CPU doing nothing.

The solution is to set the value of SurfaceView::_renderedSource to
false only in case we expect SurfaceView::paint() will be called
(visible() returns true).

Signed-off-by: Peter Tworek <tworaz666@gmail.com>